backend/requirements.txt
fastapi==0.115.8
google-genai==1.3.0
pillow==11.1.0
python-dotenv==1.0.1
uvicorn==0.34.0
vtracer==0.6.11
